requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
anyio==3.6.1
appdirs==1.4.4
asgiref==3.5.2
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
asyncer==0.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.22.5
coloredlogs==15.0.1
cycler==0.11.0
decorator==5.1.1
dlib==19.24.0
dnspython==2.2.1
email-validator==1.2.1
executing==0.8.3
face-alignment==1.3.5
fastapi==0.87.0
filelock==3.7.1
filetype==1.2.0
flatbuffers==1.12
fonttools==4.33.3
frozenlist==1.3.0
gast==0.4.0
gdown==4.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h11==0.13.0
h5py==3.7.0
httptools==0.4.0
humanfriendly==10.0
idna==3.3
ImageHash==4.3.1
imageio==2.19.3
importlib-metadata==4.11.4
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
kombu==5.2.4
libclang==14.0.1
llvmlite==0.38.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
motor==3.0.0
mpmath==1.2.1
multidict==6.0.2
networkx==2.8.3
nltk==3.7
numba==0.55.2
numpy==1.21.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.0
onnxruntime==1.13.1
opencv-python==4.5.5.64
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
orjson==3.7.1
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pooch==1.6.0
prompt-toolkit==3.0.29
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pydub==0.25.1
Pygments==2.12.0
PyMatting==1.1.8
pymongo==4.1.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.6.2
rembg==2.0.30
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.3
scipy==1.9.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
starlette==0.21.0
sympy==1.11.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-docs==0.0.0.dev0
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tifffile==2022.5.4
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.2.2.post1
typing-extensions==4.2.0
ujson==5.3.0
urllib3==1.26.9
uvicorn==0.20.0
uvloop==0.16.0
vine==5.0.0
watchdog==2.1.9
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
